requirements.txt
OpenCV
imutils
numpy
pytesseract
Pillow
streamlit